DIOS-4647: web sdk updates to eliminate circular dependency error
