+++ packages/millicast-sdk/src/utils/Diagnostics.js
@@ -1,9 +1,6 @@
-import Logger from '../Logger'
 import { version } from '../../package.json'
 
 const MAX_STATS_HISTORY_SIZE = 5
-const MAX_HISTORY_SIZE = 100
-const LOG_LEVELS = ['TRACE', 'DEBUG', 'INFO', 'WARN', 'ERROR']
 const userAgent = window?.navigator?.userAgent || 'No user agent available'
 let _accountId = ''
 let _streamName = ''
@@ -11,7 +8,6 @@ let _subscriberId = ''
 let _streamViewId = ''
 let _feedId = ''
 let _connection = ''
-let _history = []
 const _stats = []
 
 const Diagnostics = {
@@ -27,25 +23,9 @@ const Diagnostics = {
     }
     _stats.push(stats)
   },
-  get: (statsCount = MAX_STATS_HISTORY_SIZE, historySize = MAX_HISTORY_SIZE, minLogLevel = 'TRACE') => {
+  get: (statsCount = MAX_STATS_HISTORY_SIZE) => {
     if (!Number.isInteger(statsCount) || statsCount > MAX_STATS_HISTORY_SIZE || statsCount <= 0) {
       statsCount = MAX_STATS_HISTORY_SIZE
-    }
-
-    if (!Number.isInteger(historySize) || historySize > MAX_HISTORY_SIZE || historySize <= 0) {
-      throw new Error('Invalid Argument Exception : historySize must be a positive integer that is equal to or less than 100.')
-    }
-
-    _history = Logger.getHistory()
-
-    if (!LOG_LEVELS.includes(minLogLevel.toUpperCase())) {
-      throw new Error('Invalid Argument Exception : the minLogLevel parameter only excepts "trace", "debug", "info", "warn", and "error" as arguments.')
-    }
-
-    if (LOG_LEVELS.includes(minLogLevel.toUpperCase())) {
-      const filteredLogLevels = LOG_LEVELS.slice(LOG_LEVELS.indexOf(minLogLevel.toUpperCase()))
-      const filteredLevels = _history.filter((log) => filteredLogLevels.some(level => log.includes(level)))
-      _history = filteredLevels
     }
 
     const diagnostics = {
@@ -56,7 +36,6 @@ const Diagnostics = {
       streamName: _streamName,
       subscriberId: _subscriberId,
       connection: _connection,
-      history: _history.slice(-historySize),
       stats: _stats.slice(-statsCount)
     }
